+++ zaifbot/modules/indicators/moving_average.py
